linux_mysite/requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
psycopg2>=2.9,<3
django-modeltranslation>=0.18